# Optimize speed and memory for cached values returns