@AlexWayfer 2019-05-27
> Could you please provide the whole benchmark script?

I have provided. `Memery` — the file from `master`, `MemeryNew` — renamed version from this PR.