@swrobel 2019-08-03
FYI, I [benchmarked memery vs Memoist](https://gist.github.com/swrobel/fa739829d317ff5faa759c9fc5c25428#file-results). With the published gem version, memery is 1.32x slower & allocates 1.92x more. With this fix, it's roughly identical:

```
Comparison:
             memoist:   417894.3 i/s
              memery:   405734.6 i/s - same-ish: difference falls within error
             memoist:   192 allocated
              memery:   200 allocated - 1.04x more
```